+++ app/src/main/java/com/example/tvshows/database/model/PagingKeys.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class PagingKeys(
     @PrimaryKey
-    val elementId: Long,
+    val elementId: Int,
     val prevKey: Int?,
     val nextKey: Int?
 )
